+++ honey-back/ggulswriting/src/main/java/com/thirty/ggulswriting/room/repository/RoomRepository.java
@@ -8,7 +8,7 @@ import com.thirty.ggulswriting.member.entity.Member;
 import com.thirty.ggulswriting.room.entity.Room;
 
 public interface RoomRepository extends JpaRepository<Room, Integer> {
-	Optional<Room> findRoomByRoomId(Long roomId);
+	Optional<Room> findRoomByRoomId(int roomId);
 
 	Optional<Room> findRoomByMember(Member member);
 }
